+++ src/contextActions.ts
@@ -1,10 +1,10 @@
 import * as vscode from "vscode";
-import { RegisterCommandArgs } from "./types";
-import { getPath, createFile, fileWindow } from "./utils";
+import { createFile, fileWindow, getFolder } from "./utils";
 import {
   configurationChangeListener,
   initConfiguration,
 } from "./configuration";
+import { Uri } from "vscode";
 
 export const setContextActions = (context: vscode.ExtensionContext) => {
   initConfiguration();
@@ -18,14 +18,30 @@ export const setContextActions = (context: vscode.ExtensionContext) => {
 const setContextAction = (context: vscode.ExtensionContext, name: string) => {
   const command = vscode.commands.registerCommand(
     `csharpstretch.create${name}`,
-    async (options: RegisterCommandArgs) => {
-      const fileFolder = getPath(options);
-      const filePath = await fileWindow(name);
-      if (filePath === undefined) {
-        vscode.window.showErrorMessage("Could not find file path!");
-        return;
+    async (options: Uri) => {
+      if (!options) {
+        const originalClipboard = await vscode.env.clipboard.readText();
+        await vscode.commands.executeCommand("copyFilePath");
+        const folder = await vscode.env.clipboard.readText();
+        await vscode.env.clipboard.writeText(originalClipboard);
+        const newUri = await vscode.Uri.file(folder);
+        const fileFolder = getFolder(newUri);
+
+        const filePath = await fileWindow(name);
+        if (filePath === undefined) {
+          vscode.window.showErrorMessage("Could not find file path!");
+          return;
+        }
+        await createFile(fileFolder, filePath, name);
+      } else {
+        const fileFolder = getFolder(options);
+        const filePath = await fileWindow(name);
+        if (filePath === undefined) {
+          vscode.window.showErrorMessage("Could not find file path!");
+          return;
+        }
+        await createFile(fileFolder, filePath, name);
       }
-      await createFile(fileFolder, filePath, name);
     }
   );
   context.subscriptions.push(command);