+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.1.0
+
+- When creating new file with .cs extension you will get same options to create file
+- Can now create new file from command pallette and insert file to current location
+- Option to show or hide context menu on files
+
 ## 1.0.1
 
 - New C# Class Template bug fix